Unit tests

+++ db/migrate/20231010004050_create_foods.rb
@@ -4,7 +4,7 @@ class CreateFoods < ActiveRecord::Migration[7.0]
       t.string :name
       t.string :measurement_unit
       t.decimal :price, default: 0
-      t.integer :quantity, default: 0
+      t.decimal :quantity, default: 0
       t.references :user, null: false, foreign_key: { to_table: :users }
 
       t.timestamps
